refactoring ThemeContext and its usage, adding useTheme hook
