+++ src/context/ThemeContext.tsx
@@ -1,0 +1,47 @@
+import React, {createContext, useCallback, useContext} from 'react';
+import {usePersistedState} from 'src/hook/usePersistedState';
+
+type Theme = 'light' | 'dark';
+
+type ThemeContextValue = {
+  theme: Theme;
+  toggleTheme: () => void;
+};
+
+const ThemeContext = createContext<ThemeContextValue>({
+  theme: 'light',
+  toggleTheme: () => undefined,
+});
+
+type PropTypes = {
+  children: React.ReactNode;
+};
+
+export default function ThemeProvider({children}: PropTypes) {
+  const [theme, setTheme] = usePersistedState<Theme>('theme', 'light');
+
+  const toggleTheme = useCallback(() => {
+    setTheme(theme === 'dark' ? 'dark' : 'light');
+  }, [theme, setTheme]);
+
+  if (!theme) {
+    return null;
+  }
+
+  const value = {
+    theme,
+    toggleTheme,
+  };
+
+  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>;
+}
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+
+  return context;
+}
